Moved AudioFocusDurationHint to android specific file.
